advice_label_format_converter/requirements.txt
grpcio >= 1.37.0
grpcio-tools >= 1.37.0
protobuf >= 3.15.8
tqdm >= 4.62.3